+++ src/vite-bundle/src/Asset/ViteAssetVersionStrategy.php
@@ -2,7 +2,6 @@
 
 namespace Pentatrion\ViteBundle\Asset;
 
-use Pentatrion\ViteBundle\DependencyInjection\PentatrionViteExtension;
 use Pentatrion\ViteBundle\Service\FileAccessor;
 use Symfony\Component\Asset\Exception\AssetNotFoundException;
 use Symfony\Component\Asset\VersionStrategy\VersionStrategyInterface;
@@ -11,23 +10,17 @@ use Symfony\Component\HttpFoundation\RequestStack;
 /**
  * @phpstan-import-type EntryPointsFile from FileAccessor
  * @phpstan-import-type ManifestFile from FileAccessor
- * @phpstan-import-type ViteConfigs from PentatrionViteExtension
  */
 class ViteAssetVersionStrategy implements VersionStrategyInterface
 {
     private ?string $viteMode = null;
-    private string $basePath;
     /** @var ManifestFile|null */
     private ?array $manifestData = null;
     /** @var EntryPointsFile */
     private array $entrypointsData;
 
-    /**
-     * @param ViteConfigs $configs
-     */
     public function __construct(
         private FileAccessor $fileAccessor,
-        private array $configs,
         private string $configName,
         private bool $useAbsoluteUrl,
         private ?RequestStack $requestStack = null,
@@ -40,7 +33,6 @@ class ViteAssetVersionStrategy implements VersionStrategyInterface
     {
         $this->viteMode = null;
         $this->configName = $configName;
-        $this->basePath = $this->configs[$configName]['base'];
     }
 
     /**
@@ -79,7 +71,7 @@ class ViteAssetVersionStrategy implements VersionStrategyInterface
 
         if ('build' === $this->viteMode) {
             if (isset($this->manifestData[$path])) {
-                return $this->completeURL($this->basePath.$this->manifestData[$path]['file']);
+                return $this->completeURL($this->entrypointsData['base'].$this->manifestData[$path]['file']);
             }
         } else {
             return $this->entrypointsData['viteServer'].$this->entrypointsData['base'].$path;
@@ -95,7 +87,7 @@ class ViteAssetVersionStrategy implements VersionStrategyInterface
             throw new AssetNotFoundException($message, $alternatives);
         }
 
-        return $this->basePath.$path;
+        return $this->entrypointsData['base'].$path;
     }
 
     /**