+++ src/vite-bundle/src/Service/EntrypointRenderer.php
@@ -120,7 +120,7 @@ class EntrypointRenderer implements ResetInterface
         $tagRenderer = $this->getTagRenderer($configName);
 
         if (!$entrypointsLookup->hasFile()) {
-            return '';
+            return $toString ? '' : [];
         }
 
         $useAbsoluteUrl = $this->shouldUseAbsoluteURL($options, $configName);
@@ -150,6 +150,22 @@ class EntrypointRenderer implements ResetInterface
             $entrypointsLookup->isLegacyPluginEnabled()
             && !isset($this->returnedViteLegacyScripts[$configName])
         ) {
+            if ($entrypointsLookup->hasModernPolyfillsEntry()) {
+                foreach ($entrypointsLookup->getJSFiles('polyfills') as $url) {
+                    // normally only one js file
+                    $tags[] = $tagRenderer->createScriptTag(
+                        [
+                            'type' => 'module',
+                            'crossorigin' => true,
+                            'src' => $this->completeURL($url, $useAbsoluteUrl),
+                        ],
+                        '',
+                        $entryName,
+                        true,
+                    );
+                }
+            }
+
             /* legacy section when vite server is inactive */
             $tags[] = $tagRenderer->createDetectModernBrowserScript();
             $tags[] = $tagRenderer->createDynamicFallbackScript();
@@ -239,7 +255,7 @@ class EntrypointRenderer implements ResetInterface
         $tagRenderer = $this->getTagRenderer($configName);
 
         if (!$entrypointsLookup->hasFile()) {
-            return '';
+            return $toString ? '' : [];
         }
 
         $useAbsoluteUrl = $this->shouldUseAbsoluteURL($options, $configName);