+++ src/vite-plugin-symfony/src/entrypoints/entryPointsHelper.ts
@@ -65,10 +65,30 @@ export const getBuildEntryPoints = (generatedFiles: GeneratedFiles, viteConfig: 
     );
   }
 
-  if (hasLegacyEntryPoint && getOutputPath("vite/legacy-polyfills")) {
-    const fileInfos = generatedFiles[getOutputPath("vite/legacy-polyfills")!] ?? null;
+  /**
+   * legacy polyfills target browsers that don't support ESM at all.
+   * added as <script nomodule> tags for legacy browsers.
+   * @see https://github.com/vitejs/vite/tree/main/packages/plugin-legacy#polyfills
+   * */
+  const polyfills = getOutputPath("vite/legacy-polyfills-legacy");
+
+  if (hasLegacyEntryPoint && polyfills) {
+    const fileInfos = generatedFiles[polyfills];
     if (fileInfos) {
       entryPoints["polyfills-legacy"] = resolveBuildEntrypoint(fileInfos, generatedFiles, viteConfig, false);
+    }
+  }
+
+  /**
+   * modern polyfills target browsers that support ESM but lack certain widely-available features.
+   * added as <script type="module"> tags in HTML.
+   * @see https://github.com/vitejs/vite/tree/main/packages/plugin-legacy#modernpolyfills */
+  const modernPolyfills = getOutputPath("vite/legacy-polyfills");
+
+  if (modernPolyfills) {
+    const fileInfos = generatedFiles[modernPolyfills];
+    if (fileInfos) {
+      entryPoints["polyfills"] = resolveBuildEntrypoint(fileInfos, generatedFiles, viteConfig, false);
     }
   }
 
@@ -99,9 +119,10 @@ export const resolveBuildEntrypoint = (
 
       const importFileInfos = generatedFiles[importOutputRelPath];
       if (!importFileInfos) {
-        const isExternal = config.build.rollupOptions.external
+        const external = config.build.rollupOptions.external ?? config.build.rolldownOptions?.external;
+        const isExternal = external
           ? resolveUserExternal(
-              config.build.rollupOptions.external,
+              external,
               importOutputRelPath, // use URL as id since id could not be resolved
               fileInfos.inputRelPath,
               false,