+++ src/vite-bundle/CHANGELOG.md
@@ -1,5 +1,22 @@
 # `pentatrion/vite-bundle` / `vite-plugin-symfony` Changelog
 
+## v8.2.4
+
+- Support vite 8.0.0 ([van-sprundel](https://github.com/van-sprundel))
+- Fix renderScripts/renderLinks returning string despite the toString option being set to false  ([atomicptr](https://github.com/atomicptr))
+
+## v8.2.3
+
+- Add support for Symfony 8 ([@skmedix](https://github.com/skmedix))
+
+## v8.2.2
+
+- vite-plugin-symfony: Support use modernPolyfills option in @vitejs/plugin-legacy ([@twodogwang](https://github.com/twodogwang))
+
+## v8.2.1
+
+- fix #83 asset twig method not using entrypoints baseUrl ([@micheh](https://github.com/micheh))
+
 ## v8.2.0
 
 - add vite 7 support